添加新抓取请求

+++ src/main/java/com/example/demo/crawler/model/http/weixin/AppMsg.java
@@ -21,5 +21,5 @@ public class AppMsg {
     private int copyrightStat;
     private String author;
     @JsonProperty("sendtime")
-    private int sendTime;
+    private long sendTime;
 }
